+++ webview-ui/lineage/main_lineage.ts
@@ -1,9 +1,21 @@
 import { createApp } from "vue";
-import LineageApp from "../src/LineageApp.vue";
+import LineageApp from "../src/AssetLineage.vue";
 import 'highlight.js/styles/github-dark-dimmed.css';
 import { allComponents, provideVSCodeDesignSystem } from "@vscode/webview-ui-toolkit";
+import AssetLineagePanel from '@/AssetLineage.vue';
+import PipelineLineagePanel from '@/PipelineLineage.vue';
 
-const app = createApp(LineageApp)
 provideVSCodeDesignSystem().register(allComponents);
 
-app.mount("#app");
+document.addEventListener('DOMContentLoaded', () => {
+  const appElement = document.getElementById('app');
+  const componentName = appElement?.getAttribute('data-component');
+
+  if (componentName === 'AssetLineageFlow') {
+    createApp(AssetLineagePanel).mount('#app');
+  } else if (componentName === 'PipelineLineageFlow') {
+    createApp(PipelineLineagePanel).mount('#app');
+  } else {
+    console.error('Unknown component type:', componentName);
+  }
+});
